separate search component into a different components
